Fix [Workflow] Error on attempt to select deploy step
